requirements.txt
django
djangorestframework
pyjwt
ckeditor
Pillow
